app/requirements.txt
streamlit==1.28.1
langchain==0.0.335
langchain-community==0.0.6
chromadb==0.4.15
pypdf2==3.0.1
pymupdf==1.23.8
pytesseract==0.3.10
pillow==10.0.1
opencv-python-headless==4.8.1.78
pdf2image==1.16.3
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0
python-docx==0.8.11
psutil==5.9.6
GPUtil==1.4.0
redis==5.0.1
psycopg2-binary==2.9.7
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
scikit-learn==1.3.0
faiss-cpu==1.7.4
openai==0.28.1
anthropic==0.3.11
huggingface-hub==0.17.3
tokenizers==0.14.1
accelerate==0.23.0
bitsandbytes==0.41.1
xformers==0.0.22
safetensors==0.4.0